use more bits for ancestor mutation id

+++ source/EngineInterface/Descriptions.h
@@ -384,7 +384,7 @@ struct CellDescription
     LivingState livingState = LivingState_Ready;
     int creatureId = 0;
     int mutationId = 0;
-    uint8_t ancestorMutationId = 0;
+    int ancestorMutationId = 0;
     float genomeComplexity = 0;
 
     //cell function
